Rework the way data is passed into commands.
This explictly passes the session, parameters and url variables down
into each command. In addition the URL variables are handled as a
infra struct rather than as a custom thing with implied variables.

This change doesn't use the session variable in remote end steps yet;
that's part of later changes.