Redefine creating a session to handle multiple session types
This ensures that sessions are created with the correct type
flags (e.g. with the HTTP flag set for HTTP sessions). It also updates
the capabilities processing to ensure that only the relevant
capabilities are handled for each kind of session (e.g. for BiDi we
are unlikely to have a user prompt handler).

Some state, like acceptInsecureCertificates becomes
implementation-global since even if we support multiple sessions they
need to agree on that kind of browser-level behaviour.